Write logs to stdout as well as log file
If DATASOLE_LOG_OUTPUT_PATH is set, logs were previously not written
to stdout. This change pipes logs to stdout as well as file if the
environment variable is set.

Setting DATASOLE_LOG_FORMAT=json will only write JSON logs to file, not
to stdout. Stdout logs are always pretty-printed and human readable.